splits rpc
