+++ services/director-v2/src/simcore_service_director_v2/modules/osparc_variables/api_keys_manager.py
@@ -5,14 +5,12 @@ from uuid import UUID, uuid5
 
 from aiocache import cached
 from fastapi import FastAPI
-from models_library.api_schemas_webserver import WEBSERVER_RPC_NAMESPACE
-from models_library.api_schemas_webserver.auth import ApiKeyCreate, ApiKeyGet
+from models_library.api_schemas_webserver.auth import ApiKeyGet
 from models_library.products import ProductName
 from models_library.projects_nodes_io import NodeID
-from models_library.rabbitmq_basic_types import RPCMethodName
 from models_library.services import RunID
 from models_library.users import UserID
-from pydantic import BaseModel, StrBytes, parse_obj_as
+from pydantic import BaseModel, StrBytes
 from servicelib.fastapi.app_state import SingletonInAppStateMixin
 from servicelib.rabbitmq import RabbitMQRPCClient
 from servicelib.redis import RedisClientSDK, RedisClientsManager
@@ -20,63 +18,17 @@ from settings_library.redis import RedisDatabase
 
 from ...utils.base_distributed_identifier import BaseDistributedIdentifierManager
 from ..rabbitmq import get_rabbitmq_rpc_client
+from ._rpc import (
+    create_api_key_and_secret,
+    delete_api_key_and_secret,
+    get_api_key_and_secret,
+)
 
 _NAMESPACE: Final = UUID("ce021d45-82e6-4dfe-872c-2f452cf289f8")
 
 
-def create_unique_identifier_from(*parts: Any) -> str:
+def _create_unique_identifier_from(*parts: Any) -> str:
     return f"{uuid5(_NAMESPACE, '/'.join(map(str, parts)) )}"
-
-
-#
-# RPC interface
-#
-
-
-async def create_api_key_and_secret(
-    app: FastAPI,
-    *,
-    product_name: ProductName,
-    user_id: UserID,
-    name: str,
-    expiration: timedelta | None = None,
-) -> ApiKeyGet:
-    rpc_client = get_rabbitmq_rpc_client(app)
-    result = await rpc_client.request(
-        WEBSERVER_RPC_NAMESPACE,
-        parse_obj_as(RPCMethodName, "create_api_keys"),
-        product_name=product_name,
-        user_id=user_id,
-        new=ApiKeyCreate(display_name=name, expiration=expiration),
-    )
-    return ApiKeyGet.parse_obj(result)
-
-
-async def get_api_key_and_secret(
-    app: FastAPI, *, product_name: ProductName, user_id: UserID, name: str
-) -> ApiKeyGet | None:
-    rpc_client = get_rabbitmq_rpc_client(app)
-    result: Any | None = await rpc_client.request(
-        WEBSERVER_RPC_NAMESPACE,
-        parse_obj_as(RPCMethodName, "api_key_get"),
-        product_name=product_name,
-        user_id=user_id,
-        name=name,
-    )
-    return parse_obj_as(ApiKeyGet | None, result)
-
-
-async def delete_api_key_and_secret(
-    app: FastAPI, *, product_name: ProductName, user_id: UserID, name: str
-):
-    rpc_client = get_rabbitmq_rpc_client(app)
-    await rpc_client.request(
-        WEBSERVER_RPC_NAMESPACE,
-        parse_obj_as(RPCMethodName, "delete_api_keys"),
-        product_name=product_name,
-        user_id=user_id,
-        name=name,
-    )
 
 
 #
@@ -85,7 +37,7 @@ async def delete_api_key_and_secret(
 
 
 def create_user_api_name(product_name: ProductName, user_id: UserID) -> str:
-    return f"__auto_{create_unique_identifier_from(product_name, user_id)}"
+    return f"__auto_{_create_unique_identifier_from(product_name, user_id)}"
 
 
 def _build_cache_key(fct, *_, **kwargs):
